Logger/requirements.txt
termcolor
kafka-python